refactored simple udt
