+++ packages/express/src/plugin/udt/simple/srcipts/burn.ts
